Added looping foundation
Hopefully added the logic necessary for handling looping songs seamlessly. Now encourages checking a return value for whether the song is still going or has ended. Unsure if this return value was actually implemented in the non-looping section, but I'm not going to check and have to rewrite this message again